No diff offset or length for you
Use -R for that. Will add some special cases but for inversion to work,
this needs to go.
Files load fully. Use ranges to work in sections of it.
Byte offset and length changed their single letter option to capitals.
This enables -o for specifying output file name.
data.h might be a way to pass pairs of variables around. We'll see